Extracting SwarmClient to separate class;
