+++ client/src/main/java/hudson/plugins/swarm/Candidate.java
@@ -1,0 +1,17 @@
+package hudson.plugins.swarm;
+
+/**
+ *
+ */
+public class Candidate {
+
+    final String url;
+
+    final String secret;
+
+    Candidate(String url, String secret) {
+        this.url = url;
+        this.secret = secret;
+    }
+
+}
